stt_service/requirements.txt
fastapi
uvicorn
python-multipart
kafka-python
boto3
psycopg2-binary
python-dotenv
SpeechRecognition
pydub